Fixes to product repo and product controller

+++ src/main/java/com/pluralsight/northwindtradersspringboot/controller/ProductController.java
@@ -27,7 +27,7 @@ public class ProductController {
     }
 
     @GetMapping("/{id}")
-    public ResponseEntity<Product> findProductById(@PathVariable Long id) {
+    public ResponseEntity<Product> findProductById(@PathVariable int id) {
         Product product = productService.findProductById(id);
         return ResponseEntity.ok(product);
     }
@@ -46,7 +46,7 @@ public class ProductController {
     }
 
     @DeleteMapping("/{id}")
-    public ResponseEntity<Void> deleteProduct(@PathVariable Long id) {
+    public ResponseEntity<Void> deleteProduct(@PathVariable int id) {
         productService.deleteProduct(id);
         return ResponseEntity.noContent().build();
     }
